requirements.txt
APScheduler==3.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dotenv==0.0.5
Flask==2.0.3
future==0.18.2
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pyserial==3.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
six==1.16.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
Werkzeug==2.0.3
